One problem fixed, two more appeared
Fixed an issue with the AI controller repossessing an enemy after we unpossess them. Now there is an issue with the AI controls after we unpossess them. When I unpossess an astronaut, the other AIs continue to think that the attack target is the previously possessed enemy AI, not me. And the enemy that I did possess cannot seem to find the attack target.